pyproject.toml
[tool.poetry]
name = "vancelle"
version = "0.1.0"
description = ""
authors = ["Sam Clements <sam@borntyping.co.uk>"]
license = "Mozilla Public License Version 2.0"
readme = "README.md"
packages = [{ include = "vancelle", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.0"
beautifulsoup4 = "^4.12.2"
flask-alembic = "^3.0.1"
flask-cors = "^4.0.0"
flask-debugtoolbar = "^0.14.1"
flask-injector = "^0.15.0"
flask-login = "^0.6.3"
flask-sqlalchemy-lite = "^0.1.0"
flask-wtf = "^1.2.1"
frozendict = "^2.4.1"
hishel = { extras = ["sqlite"], version = "^0.0.24" }
hotmetal = "^1.0.0"
httpx = "^0.27.0"
humanize = "^4.9.0"
inflect = "^7.0.0"
platformdirs = "^4.1.0"
psycopg = { extras = ["binary"], version = "^3.1.14" }
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
setuptools = "^69.0.2"
sqlalchemy = "^2.0.23"
structlog = "^23.2.0"
svcs = "^24.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.7.1"
pytest = "^7.4.2"
types-requests = "^2.31.0.10"

[tool.poetry.group.server.dependencies]
gunicorn = "^21.2.0"

[tool.black]
line-length = 128

[tool.mypy]
namespace_packages = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=vancelle/migrations"
filterwarnings = [
    #    "ignore:The '__version__' attribute is deprecated:DeprecationWarning",
]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"
